Ride_Scanner/ride-automation-server/requirements.txt
# ride-automation-server/requirements.txt
flask==2.3.3
appium-python-client==3.1.0
selenium==4.15.0
requests==2.31.0
